Refactors `MatterPhysics` method to separate files Updates scope access for onAfterUpdate
